HETEMLMonitor/requirements.txt
paramiko==3.4.0
schedule==1.2.0
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
